Amythortz now have zero value (#2608)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Sets Amythortz Value to 0.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Amythortz having value was used by bandit mages to cheese favors, much
to my depression. This is why we can't have nice things.

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)

![image](https://github.com/user-attachments/assets/77f7ecc1-9e0c-438e-be60-73902e27db5a)

<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->