Adds area restrictions to hunting bait utility. (#2610)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Like fishing however this adds an area check for if a location is a
valid hunting spot. Currently it only checks the bog.
```
#define VALID_HUNTING_AREAS list(\
	/area/rogue/outdoors/bog )
```
Uses the check when deploying the bait.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Currently there are no restrictions to bait usage. You can deploy it
anywhere and catch animals even in high town should the rest of the
requirements succeed. It seems a bit strange that you could hunt for
goats in a graveyard no?

If it's too harsh or if you want people to deploy the bait and have it
do nothing let me know and I could do that. I just thought to include a
message to give a subtle hint towards how it should be done. At the
least it will force a couple more people into the bog for 'fun'.

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)


https://github.com/user-attachments/assets/b58899f5-85a9-4830-9c33-09d3836a2818


<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->