Anchors fishing cages so they stay in moving rivers. (#2611)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Cages move in water, it can be solved by setting anchor value to one and
back to 0 when picking up.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
See video.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)


https://github.com/user-attachments/assets/563ca48a-2275-4b35-b0b5-b42be422321f


<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->