Fixes steamy bowl sprite remaining despite being empty. (#2614)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
The bowls wouldn't update leaving you with a steamy empty bowl. Not the
biggest deal in the world but this removes it by also resetting the icon
state when it tries to update_icon
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
graphical issue, sorry the video makes it hard to see but it's there.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)


https://github.com/user-attachments/assets/49308454-113e-486e-9a4e-18eaa160e548


<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->