Amended no nervelock account message (#2619)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Tiny PR which simply adjusts the text displayed when a nervelock
attempts to deposit money to a player that doesn't have a nervelock
account. A comment in PR #2234 indicates that the name "Nervelock" was
supposed to be a replacement for the old name, but this text seems to
have been overlooked.

I feel kind of guilty about making a PR _just_ for this, but I prefer
not lumping together lots of disparate changes, and it seems like
something that _ought_ to be done but didn't have any foreseeable larger
piece of work to be attached to.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

Consistency of naming, and removal of a potentially offensive reference.

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

## Proof of Testing (Required)


![image](https://github.com/user-attachments/assets/6426ff96-9ef7-4a09-96d5-414029b6758a)

![image](https://github.com/user-attachments/assets/4e2a7b9d-debf-4b3c-8c0e-e03b5436d191)

<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->