Makes wooden forks craftable. (#2617)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Adds a datum recipe for forks.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Now that forks are usable may as well let people make them to use them.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)


https://github.com/user-attachments/assets/183fe294-493c-43f5-9444-3f23923ad430


<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->