Dye bin: fix multiple insertion bug, updates window GUI. (#2631)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Dye bins did not check for contents before inserting items. This adds
the check as well as changes the GUI for the window to also update the
color when selected (image below)
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Dye bins confuse people when they allow multiple insertions because it
only treats the most recent insertion as the only thing in it. This
prevents that. The GUI color didn't update because it was using an ascii
symbol that didn't respect the color change so now it's more obvious
what you selected.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)

![dyebin_01](https://github.com/user-attachments/assets/b4745f1f-4bdb-476e-b2d0-9801bf945078)

![dyebin_02](https://github.com/user-attachments/assets/b2f8559d-a84a-4575-bf53-0cb59a01acd5)

<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->