Fixes head attachment surgery disconnects. (#2625)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Stops player from disconnecting when head is reattached. Will it break
other things? Maybe! I only saw one runtime and that's what the other
file fixes. If this seems glaringly messed up or you know this code is
breaking something let me know.

The runtime I was getting was about getting turf lum counts but I think
it only was an issue from attaching the head to the surgery but this
could potentially be breaking how mob_movement can get a turf?? It's the
only proc that was giving me this issue so maybe the way that specific
proc tries to call turf is messed up? Not sure...

Since the brainmob thing isn't used the issue was caused by it nulling
out the already existing player's ckey in the body.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
people get disconnected and sent back to lobby when a physician tries to
save them. it's ass. please test merge ASAP
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)


https://github.com/user-attachments/assets/bce837bb-8d5f-4a86-9d8d-9e56d2c2dc0b


<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->