Bug: pot smelting, Balance: wizard robe requires silk. (#2645)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

- Removes stone pots being able to smelt into iron
- Changes wizard robe to need silk instead of cloth

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

- Pot is stone it shouldn't turn into iron ingots.
- Wizard robe is 100 coins it should need some kind of special fabrics.

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)


https://github.com/user-attachments/assets/ad268fde-715c-4aa6-b6bb-9a5c8f41e339


<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->