Stops outsiders from being announced via Scoms (#2644)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Stops migrants, goblins, bandits from being announced over scom.
Had a bit of trouble finding the right thing to use but there was a
somewhat simple solution: `peopleknowme` This is one stat that (perhaps
a little busted?) is consistent in towners that foreigners do not get.
`why the fuck did you use this??`
By that point in the code there's no way to access their job type. I
guess if this is too hacky I could try moving it around and seeing if I
can get the announcement and a proper system working. Otherwise I might
have to make a big define with all the datums of outfits to check.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
People don't wanna know when a goblin has 'arrived from kingsfield'
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)
Video is a little fast but hopefully shows it well enough. 


https://github.com/user-attachments/assets/49e08071-e4b5-4068-9a47-eb21f966604d


<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->