Gives turfs their ImmediateCalculateAdjacentTurfs() proc on init (#2606)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Gives the ```ImmediateCalculateAdjacentTurfs()``` proc back to turfs.
Seems like whatever ripped out atmos forgot about this. This proc seems
to have almost no impact on loading time either so it should be a
trivial to allow it to run on init.

If someone thinks I should rewrite smoke to not use this as there's some
other function making a list of adjacent tiles let me know and I could
try changing the smoke procs to use that instead but this was just the
easiest solution for what I wanted.

Not sure what else uses this but it would possibly fix them too as
before this whole list was always null.

**EDIT**: HOWEVER if you make a new turf it calls this proc anyway so
smoke spreads on say stone floors or twig floors you make but nothing
else by default! Messed up!
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Turfs knowing their adjacent turfs is useful for many things but it
would also, and more importantly what I wanted to do, is fix smoke. It
currently does not spread. This returns that functionality back to
smoke.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)

![smoke_fix](https://github.com/user-attachments/assets/700b8c79-7aa0-4ec3-a990-9effe9cec95c)

<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->