Removes mog from the filter (#2594)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
See title
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
There's one guy called Mog Hog which is a valid name and not a reference
to mogging and it should be an admin issue and very unimersive that we
can't call him Mog hog and call him degrade hog instead
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)
I'm actually gonna rope myself if this doesn't work
<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->