Physician Rebalancing, Adds Farming to Scholar/Physician (#2555)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
This adds farming 2 to scholar and town physician to match court
physician and the refugee alchemist role.

UPDATE:
Skills Changed - 
Sewing 3 -> 1
Crafting 2 -> 1
Carpentry 2 -> 1
Lumberjacking removed.
Added sleepcloth to bag.
Added mortar to bag.

## Why It's Good For The Game
Currently the biggest and worst thing to have as scholar/physician if
you are doing alchemy and growing your ingredients is the 20-30 minute
in-game grind just to get to farming 2. Not to mention by the time you
actually get to the point of getting all your ingredients ready the
round is mostly over since if you are working on several different
potions or brews, you still have to wait for them to brew, to grind
them, etc. Sure, you can shorten this by getting more alembics but the
wait time isn't shortened that much... It is convenient during brewing,
however.

Those of us who have tons of hours in trying to compete with the damned
merchant in selling lessers (or the steward) in making our high quality
potions just don't bother half the time or grind painfully to get
somewhere decent where in the end, it's often too late since people are
stocked up. It's a slog and honestly having something like this will
help greatly.

Court physician already gets farming 2 as well as the alchemist refugee
(which should just be removed) anyway. Physician towner house and
scholar cabinet has the means to farm but they lack the skill, unlike
the court physician. This PR will make these two roles more in-line to
the aforementioned and eliminate a lot of needless grinding, so they can
focus on selling/doing their physician/teaching part of their job if
that is the route the player wants to go down.

## Proof of Testing (Required)

![farm](https://github.com/user-attachments/assets/44714152-ee35-45bd-acf9-d604eba2e81f)