Turns the onager down a bit. (#2652)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
lowers the volumes of the more loud sounds of the onager down a bit.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
That dang onager is too loud!
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)
It's just volume controls but it should hurt people's ears less this
way.
<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->