Makes mail notifcations more obvious, adds sound. (#2651)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Reworks how you get notified about a letter (see video)
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
It gets lost in the sea of messages easily and the noise might help
people realize more.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)


https://github.com/user-attachments/assets/9869844d-0491-41d7-a47f-c99f32998732


<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->