Removes unequip bitflag; Uses existing variable for cursed collar. (#2649)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Turns out there is an existing method for checking if something can be
removed from yourself that is functional. I made the cursed collar use
that and removed the bitflag logic for the time being. There's some
other items that might be needing this too so I'll probably make a
future PR to help with what are otherwise broken items.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Consolidates the methods for stopping unequip. Listing the var by the
bitflag too in case someone wonders why it's commented out.

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)
Example of cursed collar using the preexisting var, same functionality
as with the bit flag.



https://github.com/user-attachments/assets/39a6341b-eb60-43dc-a3b8-6c647b39a8b6


<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->