NewMoon mercenary buffs and changes (#2660)
## About The Pull Request

More discussion in Issue #2657

New Moon Spellblades are bad. They are worse than the Battlemagos. A
refugee is usually a good bit worse than the corresponding mercenary.
This PR buffs the mageskills and items of newmoon mercenaries, nerfs
their other skills a bit. Buffs Con stat.

In somewhat more Detail:
They start with six instead of three spellpoints, and have
prestidigitation by default.
They have a lesser manapot in their backpack, mercenaries should not be
by default encouraged to bogtrek as much.
They have a beginner spellbook instead of a horrible one.
They have increased reading (1->2)
They have increased arcane (2->3)

## Why It's Good For The Game

Moves the Newmoon closer to their scholarly origin. Puts their balance
better in line with Mercenaries being superior to refugees generally.
All numbers and such here are completely temporary, I just want to start
a discussion with balance maintainers regarding this unique merc class.


Many other things come to mind which would have to be atomized but can
be discussed here in this PR, such as adding a bit of chalk, cinnabar or
manapots to the mercenary guild aswell as buffing the cool frost sabre
and or their jacket.


## Proof of Testing (Required)

I ran it in a local server and it seemed to work out as expected.


![IkFROOO](https://github.com/user-attachments/assets/d89c161d-eb8b-42e4-816b-819aaebeafd2)