Adds a handful of descriptors for character creation (#2627)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
This PR is mainly for me to figure out how the workflow is like for
github. It adds a few new descriptors to use for character creation.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Mostly unnecessary... It's just for flavour and to let me figure out how
to make PRs.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)
![descriptors
proof](https://github.com/user-attachments/assets/99acb3d0-88bb-4ca8-b809-5eca28207477)

<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->