Fixes wake-up not removing sleep debuff. (#2642)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Why oh why did sleep change how it functions? Anyway.. This makes the
pills remove the debuff on consumption. I'd rather that be moved back to
the value but if this is the way it's going to be then this is the
solution...
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Wake-up pills have no effect on sleep debuff (they would via coffee if
`AdjustSleeping` and the var for sleepiness in the mob worked!!! Was it
changed? Why??
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)
Not shown: it working on other people ( I did test this just in case and
it does also work feeding someone else)


https://github.com/user-attachments/assets/bcb28d10-3306-4429-89de-d7445af708f3


<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->