Make Test merge links open in a new browser window in Byond 516. (#2653)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

Currently, pressing a test merge link upon joining replaces the goonchat
page in the game with a Github login, losing your current chat history.
Gross.

Fixes that so it opens in a new browser window per old behavior. The
solution uses `usr << link()` so it should work with all supported
client versions.

## Why It's Good For The Game

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

People like their chat to not be replaced with broken web pages.

## Proof of Testing (Required)

Awkward to test as I had to mock in imaginary test merge data, but I'm
not testing TGS. The screenshot opens a random issue as I just typed in
1234 as the PR number.


![image](https://github.com/user-attachments/assets/fa565543-7a18-44ca-88a9-ceb607c9aeaf)

<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->