Knee stinger adjustments and quality of life (#2592)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
All knee stingers have been removed, and knee stingers have been
manually placed based on the design idea of where they were in the old
sewer.
A lot of bog areas have been given their proper roof, helping vampires
and werewolf stay indoors without being burned alive.
A bugfix was made for the right side of the bog, where another layer of
grass and trees was made, where they were not supposed to be.
Knee stingers were removed in the river as per request.
Spiders under the city were replaced with rats since some houses (and
the utility shop) had constant spider sounds all game.
Requested changes to remove horses were removed from the keep.
Moved the carriage to the city, and gave it a little area to be in.
Requested change to remove two Saiga from the vanguard stable.
Added a missing butcher knife to the butcher house.
Added dead trees along the road, for a content PR that are in the work.
Deleted the leftover fog bandit wall.
Changed the direction of the vanguard doors to be put on the right way.
Artificer shutters
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
The entirety of the sewer was filled with 300 knee stingers, hidden in
bushes, under the iron bars (for some reason) I have no idea why we have
this hostile of a map made for the sewers, and the druid area, but it
made the entirety of the area not useable for anyone that did not have
protection.

Instead, I put knee stingers based on the old map, where they are put in
"plain sight" so anyone that reacts to them, and not get randomly
stunned for something that was hidden in a bush.
Amount with the old map

![image](https://github.com/user-attachments/assets/693a9988-5510-4e8f-866c-d9ddd2a79273)
New amount

![image](https://github.com/user-attachments/assets/f84a2e4b-51f8-498e-8d07-82c07b027115)
Previously

![image](https://github.com/user-attachments/assets/0fd300ab-5226-4bd1-bce1-b11dbd465078)


<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)
Video of the changes, so people can see where the new knee stingers are
(hopefully, I got them all)
https://youtu.be/D_azMQnBz5o

Requested Saiga removal and carriage being moved to the town
https://youtu.be/JavIe2bBI58

Dead trees along the road
https://youtu.be/sQI9fxh2Hmw

New artificer shutters

![image](https://github.com/user-attachments/assets/52fd37fb-501d-4354-8625-51040eb69037)

![image](https://github.com/user-attachments/assets/2fdf3394-4807-49b1-b175-7f62c0b800b4)

Horses in the vanguard

![image](https://github.com/user-attachments/assets/ad08ceb9-572b-491d-a3c4-9d787c526461)


<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->