Mating Press, Part 2 (#2661)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Built from Mating Press #2609, credit to @Dilligentdog 

### **See original PR for further details.**

* Force mating press to pin target to ground
* Force move user to target after pin
* Addition of "on_start" function
* Alteration of pain stats for GOODLOVER to match normal Tweaking of
flavor text
* Removed fetish verbs armpit and foot lick
* Action will cause the TARGET to also take stamina damage, something
unique to this verb (but at a much reduced rate than the USER)

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

Adds some unique behavior and a more brutal style of fucking someone.
Variety is the spice of life!

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

## Proof of Testing (Required)


![MP1](https://github.com/user-attachments/assets/9be94dc1-07a8-4634-ba4e-479bdeb89207)
This action requires a grab, and will force the target to the ground,
moving the user over the target as shown.
I dont believe the action to be meta, even if its 100% chance
force-down, as it requires both parties to be naked.


![MP2](https://github.com/user-attachments/assets/37bca09e-dc20-4ff2-8e7b-529ee009ebd4)
Proof that without a STRONG grab, you cannot mating press.


NOTE: Foot and Armpit Lick were removed by request to make room on the
verb menu for this action.

<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->