Stops Butterflies from Destroying Seeds (#2665)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
First PR. Fix to stop butterflies from destroying seeds for latejoin
Druids. Changes seeds to only allow HUMAN or larger mobs to destroy them
when stepped on.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Latejoin Druids will have seeds to spawn with. The damn butterflies will
no longer be a menace.

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)


https://github.com/user-attachments/assets/d222282b-d538-4a7d-9611-c9ae0a56a07a


<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->