Throne command to Invalidate decrees (#2671)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Adds new command to remove old decrees
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Currently there is no way to remove old decrees and it is bad
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)

![image](https://github.com/user-attachments/assets/3cba145e-9040-4325-8b6c-08b71527bda7)

![image](https://github.com/user-attachments/assets/43d35682-d9f7-49ac-ba60-045e6814f71a)

![image](https://github.com/user-attachments/assets/1e553801-19b4-4653-9859-eed6aace11cb)

![image](https://github.com/user-attachments/assets/81360664-bf28-447c-9d81-5e1f2eb499b5)

<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->