Adds icon to examine text of atoms. (#2672)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Adds a little icon to the text box
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
I just think it's neat. Also might help a little in figuring out what
you examine?
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)

![image](https://github.com/user-attachments/assets/6715ce6d-62ef-4038-ad8d-0fff780a8c65)

Increases with the font size too!

![image](https://github.com/user-attachments/assets/b6317c43-6401-4d22-adc0-4ec5c0872870)

<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->