Blacksteel Expansion (#2624)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
This PR adds blacksteel halfplates,
Blacksteel hammers, blacksteel picks, and blacksteel tongs.
Tongs are mostly cosmetic with a 2 defense stat.
The hammer has a better hammer quality for smithing, and functions with
chisels.
Pickaxes have more integrity.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
More blacksteel options are good, if not for formal use, for the
prestige of having it. More flavor and what not
Someone sprited these months back and others were supposed to add it-
but never got around to it I guess.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->


## Proof of Testing (Required)

![image](https://github.com/user-attachments/assets/63182abe-64b9-47e1-9f9a-3f8b21d0dbc8)

![image](https://github.com/user-attachments/assets/dcd7fcb9-30a9-4fbb-a364-c50323e005c9)

<!-- Show proof of testing, screenshots or recordings for features,
proof of compile for backend changes -->